# swagger.json
{"openapi":"3.0.0","paths":{"/v1.0/user-management/user":{"get":{"operationId":"UserManagementController_getUserInfo","summary":"Get the user details","description":"Get the user details","deprecated":false,"parameters":[{"name":"organization-id","required":true,"in":"query","example":"a92d0bda-6574-4e62-a54b-315bcaaf2456","schema":{}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/JsendResponseSuccess"},{"type":"object"},{"properties":{"data":{"type":"object","$ref":"#/components/schemas/DmproUserDtoWithDates"}}}]}}}}},"tags":["User-management"],"security":[{"bearer":[]}]}}},"info":{"title":"user/organization management","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthRequestDto":{"type":"object","properties":{}},"JsendResponseSuccess":{"type":"object","properties":{"status":{"type":"string","description":"API Call response status","example":"success"},"data":{"type":"object","description":"Data"}},"required":["status","data"]},"UserImageUrl":{"type":"object","properties":{"url":{"type":"string","description":"user imageURL","example":"https://cdncvr.duclo.net/fd121cea-2610-42f0-be75-8cf508bee221/accounts/e047ac6a-b763-41a3-95bb-017d5ebdb62c/profile-1694773305544.png?Policy=eyJTdGF0ZW1lbnQiOiBbeyJSZXNvdXJjZSI6Imh0dHBzOi8vY2RuY3ZyLmR1Y2xvLm5ldC9mZDEyMWNlYS0yNjEwLTQyZjAtYmU3NS04Y2Y1MDhiZWUyMjEvYWNjb3VudHMvZTA0N2FjNmEtYjc2My00M1EzLTk1YmItMDE3ZDVlYmRiNjJjL3Byb2ZpbGUtMTY5NDc3MzMwNTU0NC5wbmciLCJDb25kaXRpb24iOnsiRGF0ZUxlc3NUaGFuIjp7IkFXUzpFcG9jaFRpbWUiOjE2OTUxNzE2MjZ9fX1dfQ__&Signature=G9MafWIMyNMuKouTmce90qaM8t-er8wC8jWtIAFC3apuVIankgcBvWf1S9gWcBNS95lJi2g2aogENyUVJoKRSCXw-XnWcDBEoLbCf01Rs13OksiKZBSNP~8sUSvxD83aYsmvHlisfB5WOAwgujHogPkdL6NJykGA4lR1VeW766YV6BnSYo2qD3Oh0DFq00ztPj~zsMCJoldf-OeirjsTtcg~sRm2fyDYtCnramJDzzpfF89IeFbl7jgzfxE98~NgwZx81~MuPhOYEf6zfcC-o~p-fldjbAD1nrhD66~W6UJHKxO0w~oAIlwdp0HEbgnXHYxtFOZaHa-GYZ4AmBQRfQ__&Key-Pair-Id=APKA2F4WRZOGAT47BO76"},"expiry":{"type":"string","description":"expiry","example":"1695171626000"}},"required":["url","expiry"]},"OrganizationDto":{"type":"object","properties":{"orgId":{"type":"string","description":"OrganizationId","example":"b45f4f01-003d-4152-b752-945cf3be0785"},"orgName":{"type":"string","description":"OrganizationName","example":"DMPro Customer Organization"},"orgType":{"type":"string","description":"OrganizationType","example":"CUSTOMER_ORG"},"addressLine1":{"type":"string","description":"Address line 1","example":"6 Pangyo-ro 319beon-gil, Bundang-gu, Seongnam-si, Gyeonggi-do, South Korea"},"addressLine2":{"type":"string","description":"Address line 2","example":""},"city":{"type":"string","description":"city","example":"Seongnam-si"},"stateName":{"type":"string","description":"stateName","example":"Gyeonggi-do"},"country":{"type":"string","description":"country","example":"South Korea"}},"required":["orgId","orgName","orgType","addressLine1","addressLine2","city","stateName","country"]},"LicenseDto":{"type":"object","properties":{"licenseId":{"type":"string","description":"License Id","example":"39f35a25-ebd6-499f-8e44-d58e88500d8e"}},"required":["licenseId"]},"AreaDto":{"type":"object","properties":{"areaId":{"type":"string","description":"Area Id","example":"3cec441b-d46d-4aa9-9083-b7e6d648343b"},"devices":{"description":"List of Device Id in Area","example":"[\"ca1b107d-2dbf-497f-b836-7187626569a4\", \"5fdadae3-8a69-4b77-bdae-40a327dcbeb2\"]","type":"array","items":{"type":"array"}}},"required":["areaId","devices"]},"LocationDto":{"type":"object","properties":{"locationId":{"type":"string","description":"Location Id","example":"3cec441b-d46d-4aa9-9083-b7e6d648343b"},"areas":{"description":"List of Area Information in location","type":"array","items":{"$ref":"#/components/schemas/AreaDto"}},"devices":{"description":"List of Device Id in location","example":"[\"e68477b5-20f2-4ad1-9f87-a6a3ca35f414\"]","type":"array","items":{"type":"array"}}},"required":["locationId","areas","devices"]},"DmproUserDto":{"type":"object","properties":{"userId":{"type":"string","description":"userId","example":"080d7837-1631-424b-96d3-6876f82ffc0c"},"firstName":{"type":"string","description":"firstName","example":"hwan"},"lastName":{"type":"string","description":"lastName","example":"kim"},"email":{"type":"string","description":"email","example":"hwan.kim@hanwha.com"},"accountLanguage":{"type":"string","description":"language","example":"en_US"},"phoneNumber":{"type":"string","description":"language","example":"en_US"},"accountStatus":{"type":"string","description":"accountStatus","example":"PENDING_ACTIVATION"},"portalRole":{"type":"string","description":"role","example":"Admin"},"dmproRole":{"type":"string","description":"DM Pro Role","example":"manager | engineer"},"dmproClaimed":{"type":"boolean","description":"User Claimed to DM Pro","example":"true"},"isLocked":{"type":"boolean","description":"isLocked","example":true},"imageUrl":{"description":"user image","allOf":[{"$ref":"#/components/schemas/UserImageUrl"}]}},"required":["userId","firstName","lastName","email","accountLanguage","accountStatus","portalRole","dmproRole","dmproClaimed","isLocked","imageUrl"]},"DmproTokenValidation":{"type":"object","properties":{"validAccessToken":{"type":"boolean","description":"category for personalization","example":"dmpro"},"validRefreshToken":{"type":"boolean","description":"category for personalization","example":"dmpro"}},"required":["validAccessToken","validRefreshToken"]},"ResClaimUserInOrgDto":{"type":"object","properties":{"userId":{"type":"string","description":"userId","example":"080d7837-1631-424b-96d3-6876f82ffc0c"},"dmproRole":{"type":"string","description":"DM Pro Role","example":"manager | engineer"},"dmproClaimed":{"type":"boolean","description":"User Claimed to DM Pro","example":"true"}},"required":["userId","dmproRole","dmproClaimed"]},"ResUpdateUserRoleInOrgDto":{"type":"object","properties":{"userId":{"type":"string","description":"userId","example":"080d7837-1631-424b-96d3-6876f82ffc0c"},"dmproRole":{"type":"string","description":"DM Pro Role","example":"manager | engineer"}},"required":["userId","dmproRole"]},"ResUnclaimUserInOrgDto":{"type":"object","properties":{"userId":{"type":"string","description":"userId","example":"080d7837-1631-424b-96d3-6876f82ffc0c"},"dmproRole":{"type":"string","description":"DM Pro Role","example":"NULL"},"dmproClaimed":{"type":"boolean","description":"User Claimed to DM Pro","example":"false"}},"required":["userId","dmproRole","dmproClaimed"]},"DmproUserDtoWithDates":{"type":"object","properties":{"userId":{"type":"string","description":"userId","example":"080d7837-1631-424b-96d3-6876f82ffc0c"},"firstName":{"type":"string","description":"firstName","example":"hwan"},"lastName":{"type":"string","description":"lastName","example":"kim"},"email":{"type":"string","description":"email","example":"hwan.kim@hanwha.com"},"accountLanguage":{"type":"string","description":"language","example":"en_US"},"phoneNumber":{"type":"string","description":"language","example":"en_US"},"accountStatus":{"type":"string","description":"accountStatus","example":"PENDING_ACTIVATION"},"portalRole":{"type":"string","description":"role","example":"Admin"},"dmproRole":{"type":"string","description":"DM Pro Role","example":"manager | engineer"},"dmproClaimed":{"type":"boolean","description":"User Claimed to DM Pro","example":"true"},"isLocked":{"type":"boolean","description":"isLocked","example":true},"imageUrl":{"description":"user image","allOf":[{"$ref":"#/components/schemas/UserImageUrl"}]},"activeDate":{"type":"number","description":"Active Date","example":"1698364592820"},"createdDate":{"type":"number","description":"Created Date","example":"1698277376653"},"updatedDate":{"type":"number","description":"Updated date","example":"1701890984919"}},"required":["userId","firstName","lastName","email","accountLanguage","accountStatus","portalRole","dmproRole","dmproClaimed","isLocked","imageUrl","activeDate","createdDate","updatedDate"]}}}}